добавление возможности для удаления тэгов в приложении Notes

+++ Notes/app/src/main/java/ru/freeit/notes/data/repository/TagRepositoryImpl.kt
@@ -15,22 +15,22 @@ class TagRepositoryImpl(
     private val tagDao = database.tagDao()
 
     private val tagsToBeAdded = mutableListOf<Tag>()
-
-    override suspend fun remove(tag: Tag) = withContext<Unit>(dispatcher) {
-        if (tag.isBoundedToNote()) {
-            tagDao.remove(tag.toDb())
-        } else {
-            tagsToBeAdded.remove(tag)
-        }
-    }
+    private val tagsToBeRemoved = mutableListOf<Tag>()
 
     override suspend fun removeBy(noteId: Long) = withContext(dispatcher) { tagDao.removeBy(noteId) }
+    override fun remove(tag: Tag) { tagsToBeRemoved.add(tag) }
     override fun add(tag: Tag) { tagsToBeAdded.add(tag) }
     override fun tags() = tagsToBeAdded
+    override fun clear() {
+        tagsToBeAdded.clear()
+        tagsToBeRemoved.clear()
+    }
 
     override suspend fun apply(noteId: Long) = withContext(dispatcher) {
+        tagsToBeRemoved.forEach { tag ->
+            if (tag.isBoundedToNote()) { tagDao.remove(tag.toDb()) } else { tagsToBeAdded.remove(tag) }
+        }
         tagsToBeAdded.forEach { tag -> tagDao.add(tag.copy(noteId = noteId).toDb()) }
-        tagsToBeAdded.clear()
     }
 
 }
